Makes diagnostics persistent across files
I moved `diagnosticCollection` and `consoleOutput` into the functions that use them. Then I changed `parseAu3CheckOutput()` to get the URI of the document so that we can delete the diagnostics for that particular document when AU3Check returns with no warnings or errors for it. Deleting that part of the diagnosticCollection enables us to keep showing the other problems instead of losing everything with `.clear()`, which I moved to call only when diagnostics are disabled.