Consolidated symbol code
I made workspaceSymbols basically call `provideWorkspaceSymbols` (which was a better implementation) for each document and cache all of it until something changes in the workspace. Got rid of the query work since the search bar will filter over all of them anyway and we're not running through everything each time the hotkey is hit.